% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{avoid_gc}
\alias{avoid_gc}
\title{Adjust P-values to Avoid Genomic Control Artifacts}
\usage{
avoid_gc(targ, log_file = NULL)
}
\arguments{
\item{targ}{A data frame containing SNP data with Beta coefficients, standard errors, and P-values.}

\item{log_file}{Optional; path to a log file where messages will be recorded.}
}
\value{
The input data frame with adjusted P-values.
}
\description{
This function checks for and adjusts P-values that may have been affected by genomic control. It recalculates P-values using available standard errors and Beta coefficients.
}
\examples{
# Read in example clean sumstats
adjusted_data <- avoid_gc(targ = clean_sumstats_1)
}
