% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{calc_mean_freq}
\alias{calc_mean_freq}
\title{Calculate average allele frequency across cases and controls}
\usage{
calc_mean_freq(sumstats, sampling = NA, log_file = NULL)
}
\arguments{
\item{sumstats}{A data frame containing GWAS summary statistics, which should include columns for allele frequencies in cases (`FRQ_A`), controls (`FRQ_U`), and optionally the number of cases (`N_CAS`) and controls (`N_CON`).}

\item{sampling}{Optional; a numeric value representing the sampling fraction of cases in the study population. If not provided or NA, and `N_CAS` and `N_CON` are missing, the function assumes equal case and control numbers.}

\item{log_file}{Optional; a string specifying the path and name of the log file to record the process of the function. If NULL or not provided, no logging is performed.}
}
\value{
Returns a numeric vector representing the average allele frequency across cases and controls. If allele frequency information is not available in `sumstats`, it returns NULL.
}
\description{
This function calculates the average allele frequency across cases and controls in GWAS (Genome-Wide Association Study) summary statistics. It accounts for different scenarios based on the available data, such as when case and control counts are available, or when a sampling fraction is provided. The function also logs the method used for the calculation.
}
\examples{
mean_freq <- calc_mean_freq(clean_sumstats_1, sampling = 0.5)
print(mean_freq)
}
