% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat0}
\alias{cat0}
\title{Concatenate and Print Objects With an Optional Separator}
\usage{
cat0(..., sep = "")
}
\arguments{
\item{...}{Arguments passed to the base `cat` function.
These can include objects to be concatenated and printed.}

\item{sep}{A string specifying the separator to be used between the objects.
Defaults to an empty string, meaning that objects will be concatenated without any separator.}
}
\value{
Invisibly returns `NULL`. It is used primarily for its side effect of printing.
}
\description{
This function is a wrapper around the base R `cat` function, with the default separator set to an empty string. It concatenates and prints its arguments.
}
\examples{
cat0("Hello", "World") # prints "HelloWorld"
cat0("Hello", "World", sep = " ") # prints "Hello World"
cat0("Line1", "Line2", sep = "\n") # prints "Line1" and "Line2" on separate lines
}
