% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHROMS}
\alias{CHROMS}
\title{Default range for chromosome numbers}
\format{
An object of class \code{integer} of length 22.
}
\usage{
CHROMS
}
\description{
A numeric vector containing chromosome numbers 1 to 22.
}
\keyword{datasets}
