% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{detect_strand_flip}
\alias{detect_strand_flip}
\title{Detect Strand Flips in SNP Data}
\usage{
detect_strand_flip(targ, ref)
}
\arguments{
\item{targ}{A vector of IUPAC codes for the target SNPs.}

\item{ref}{A vector of IUPAC codes for the reference SNPs.}
}
\value{
A logical vector indicating which SNPs are flipped.
}
\description{
This function identifies SNPs in the target that are on the opposite strand compared to the reference.
}
\examples{
flipped_snps <- detect_strand_flip(iupac_codes$target, iupac_codes$reference)
}
