% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{discord_maf}
\alias{discord_maf}
\title{Remove Variants with Discrepancy in Reported and Reference Allele Frequencies}
\usage{
discord_maf(
  targ,
  thresh,
  log_file = NULL,
  plot_file = NA,
  max_drop_frac = 0.25
)
}
\arguments{
\item{targ}{A data frame containing SNP data with both reported and reference allele frequencies.}

\item{thresh}{A numeric threshold for the absolute frequency difference.}

\item{log_file}{Optional; path to a log file where messages will be recorded.}

\item{plot_file}{Optional; path to save a plot file visualizing the frequency discrepancies.}

\item{max_drop_frac}{Optional; Maximum fraction of variants with discordant FREQ to reference.}
}
\value{
A filtered data frame with variants within the specified frequency discrepancy threshold.
}
\description{
This function filters out variants from the target dataset where the absolute difference between reported and reference allele frequencies exceeds a specified threshold. Optionally, it can generate a plot of these discrepancies.
}
\examples{
filtered_data <- discord_maf( targ = clean_sumstats_1,
                              thresh = 0.01,
                              plot_file = paste0(tempdir(),"/plot.png"),
                              max_drop_frac = 0.9)
}
