% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{filter_info}
\alias{filter_info}
\title{Filter Variants Based on INFO Score Threshold}
\usage{
filter_info(targ, thresh, log_file = NULL)
}
\arguments{
\item{targ}{A data frame of SNP data with an INFO column.}

\item{thresh}{A numeric threshold for the INFO score.}

\item{log_file}{An optional path to a log file where messages will be recorded.}
}
\value{
A filtered data frame of SNP data.
}
\description{
This function removes variants from the target data that have an INFO score below the specified threshold.
}
\examples{
filtered_data <- filter_info(clean_sumstats_1, 0.8)
}
