% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{filter_maf}
\alias{filter_maf}
\title{Filter Variants Based on Minor Allele Frequency (MAF) Threshold}
\usage{
filter_maf(targ, thresh, ref = F, log_file = NULL)
}
\arguments{
\item{targ}{A data frame containing SNP data.}

\item{thresh}{A numeric threshold for MAF.}

\item{ref}{Logical, if TRUE, uses the `REF.FREQ` column for filtering, otherwise uses `FREQ`.}

\item{log_file}{Optional; path to a log file where messages will be recorded.}
}
\value{
A filtered data frame with variants meeting the MAF threshold.
}
\description{
This function removes variants from the target dataset that have a minor allele frequency (MAF) below a specified threshold. It uses either the reported MAF (`FREQ` column) or the reference MAF (`REF.FREQ` column), depending on the `ref` argument.
}
\examples{
filtered_data <- filter_maf(targ = clean_sumstats_1, thresh = 0.05, ref = FALSE)
}
