% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{filter_n}
\alias{filter_n}
\title{Filter Variants Based on Sample Size Deviation}
\usage{
filter_n(targ, n_sd = 3, log_file = NULL)
}
\arguments{
\item{targ}{A data frame containing SNP data with a sample size column (`N`).}

\item{n_sd}{The number of standard deviations from the median to use as the threshold.}

\item{log_file}{Optional; path to a log file where messages will be recorded.}
}
\value{
A filtered data frame with variants within the specified sample size range.
}
\description{
This function removes variants from the target dataset where the sample size (`N`) is more than a specified number of standard deviations away from the median sample size.
}
\examples{
filtered_data <- filter_n(targ = clean_sumstats_1, n_sd = 3)
}
