% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{format_header}
\alias{format_header}
\title{Check GWAS sumstat header is valid and update based on interpreted meaning}
\usage{
format_header(sumstats, log_file = NULL)
}
\arguments{
\item{sumstats}{A data frame containing GWAS summary statistics where the column names are expected to be in the header.}

\item{log_file}{An optional string specifying the path and name of the log file where messages about the header interpretation and validation will be recorded. If NULL, messages are printed to the console.}
}
\value{
A data frame of GWAS summary statistics with updated and validated column names based on the interpreted header.
}
\description{
This function checks if the header of a GWAS (Genome-Wide Association Studies) summary statistics dataset is valid and updates it based on interpreted meanings. It first interprets the header using `head_interp`, then logs the interpretation, checks for the presence of required columns, logs any errors or warnings, and finally updates the summary statistics dataset with the interpreted header.
}
\examples{
formatted_ss_data <- format_header(raw_sumstats_1)
print(formatted_ss_data)
}
