% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person to say hello to.}
}
\value{
The output from \code{\link{print}}
}
\description{
Prints 'Hello' followed by a person's name
}
\examples{
hello("John")

\dontrun{
hello("Steve")
}
}
