% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{insert_beta}
\alias{insert_beta}
\title{Convert Odds Ratios or Z-scores to Beta Coefficients}
\usage{
insert_beta(targ, log_file = NULL)
}
\arguments{
\item{targ}{A data frame containing SNP data with OR or Z-scores.}

\item{log_file}{Optional; path to a log file where messages will be recorded.}
}
\value{
The input data frame with added `BETA` column.
}
\description{
This function calculates Beta coefficients for SNP data, using either available odds ratios (OR) or Z-scores and allele frequencies, if Beta coefficients are not already present.
}
\examples{
updated_data <- insert_beta(targ = clean_sumstats_1)
}
