% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{insert_se}
\alias{insert_se}
\title{Compute Standard Error from Beta and P-value}
\usage{
insert_se(targ, log_file = NULL)
}
\arguments{
\item{targ}{A data frame containing SNP data with Beta coefficients and P-values.}

\item{log_file}{Optional; path to a log file where messages will be recorded.}
}
\value{
The input data frame with added `SE` column.
}
\description{
This function calculates the standard error (SE) for each SNP's Beta coefficient, based on its P-value, if SE is not already present.
}
\examples{
updated_data <- insert_se(targ = clean_sumstats_1)
}
