% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_add}
\alias{log_add}
\title{Add text to an existing log file}
\usage{
log_add(log_file = NULL, message, sep = "\\n")
}
\arguments{
\item{log_file}{A string specifying the path and name of the log file to be appended. If NULL (default), the message is printed to the console.}

\item{message}{A string or object containing the text message to be added to the log file.}

\item{sep}{A string specifying the separator to be used after the message. Defaults to a newline character ('\\n').}
}
\value{
Invisibly returns NULL. The function is used for its side effect of writing to a log file or printing to the console.
}
\description{
This function adds a specified text message to an existing log file. If no log file is specified, the message is printed to the console.
}
\examples{
log_add(paste0(tempdir(),"/logfile.log"), "This is a log message")
log_add(message = "This is a console message")
}
