% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_header}
\alias{log_header}
\title{Create log file with standard header including script name and command line options}
\usage{
log_header(log_file, opt, script, start.time)
}
\arguments{
\item{log_file}{A string specifying the path and name of the log file to be created or appended.}

\item{opt}{A list or other object containing the command line options to be written to the log file.}

\item{script}{A string specifying the name of the script for which the log is being created.}

\item{start.time}{The start time of the process or operation, expected to be an object of class POSIXct as obtained from Sys.time(). This parameter typically captures the time at which a particular process or analysis was initiated. It's used within the function to calculate durations, log time stamps, or for other time-related operations.}
}
\value{
Invisibly returns NULL. This function is used for its side effect of writing to a log file.
}
\description{
This function creates a log file and writes a standard header to it. The header includes the script name, contact information, and any command line options passed to the script. It's useful for tracking and documenting script runs.
}
\examples{
# Assuming 'opts' is a list of options and 'script_name' is the name of the script
start_time <- Sys.time()
log_header(paste0(tempdir(),"/logfile.log"), opt_list, "script_name.R", start_time)
}
