% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{neff}
\alias{neff}
\title{Calculate Effective Sample Size}
\usage{
neff(ncas, ncon)
}
\arguments{
\item{ncas}{The number of cases in the study.}

\item{ncon}{The number of controls in the study.}
}
\value{
Returns the effective sample size as a numeric value.
}
\description{
This function calculates the effective sample size for a study based on the number of cases and controls.
}
\examples{
effective_size <- neff(1000, 1500)
}
