% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{opt_to_df}
\alias{opt_to_df}
\title{Converts a list of parameters into a table}
\usage{
opt_to_df(opt)
}
\arguments{
\item{opt}{List of parameters}
}
\value{
data.frame with two columns called 'Parameter' and 'Value'
}
\description{
Converts a list of parameters into a table
}
\examples{
opt<-list(sumstats = 'data/sumstats.gz', population = 'EUR', n = 2309)
opt_to_df(opt)

}
