% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{remove_ambig}
\alias{remove_ambig}
\title{Remove Ambiguous SNPs from Data}
\usage{
remove_ambig(dat)
}
\arguments{
\item{dat}{A data frame containing SNP data with IUPAC codes or alleles (A1, A2).}
}
\value{
A data frame with ambiguous SNPs removed.
}
\description{
This function filters out SNPs with ambiguous IUPAC codes (R, Y, K, M) from the given data.
}
\examples{
cleaned_data <- remove_ambig(clean_sumstats_1)
}
