% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{snp_allele_comp}
\alias{snp_allele_comp}
\title{Convert SNP Alleles to Their Complementary Alleles}
\usage{
snp_allele_comp(x = NA)
}
\arguments{
\item{x}{A character vector containing alleles ('A', 'T', 'C', 'G').}
}
\value{
A character vector with the complementary alleles.
}
\description{
This function converts each allele in a given vector to its complementary allele.
}
\examples{
complementary_alleles <- snp_allele_comp(c('A', 'C', 'G', 'T'))
}
