% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_cleaner.R
\name{snp_iupac}
\alias{snp_iupac}
\title{Create IUPAC Codes for pairs of SNP alleles}
\usage{
snp_iupac(x = NA, y = NA)
}
\arguments{
\item{x}{A character vector representing the first nucleotide of the allele pairs. Possible values are 'A', 'T', 'C', 'G'.}

\item{y}{A character vector of the same length as `x`, representing the second nucleotide of the allele pairs. Possible values are 'A', 'T', 'C', 'G'.}
}
\value{
A character vector of IUPAC codes corresponding to the input SNP allele pairs. If `x` and `y` are different lengths, the function prints an error message and does not return a value.
}
\description{
This function takes two vectors representing pairs of alleles for SNP (Single Nucleotide Polymorphism) pairs and converts them into IUPAC (International Union of Pure and Applied Chemistry) codes. It returns an error message if the input vectors are of different lengths.
}
\examples{
iupac_codes <- snp_iupac(clean_sumstats_1$A1, clean_sumstats_1$A2)
}
