% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat0}
\alias{cat0}
\title{cat function with '' as default separator}
\usage{
cat0(..., sep = "", file = "", append = FALSE)
}
\arguments{
\item{...}{Arguments passed to the base 'cat' function. These can include objects to be concatenated and printed.}

\item{sep}{A string specifying the separator to be used between the objects. Defaults to an empty string.}

\item{file}{A connection, or a character string naming the file to print to. If "" (the default), cat0 prints to the standard output connection.}

\item{append}{Logical. If TRUE, the output will be appended to the file; otherwise, it will overwrite the contents of the file.}
}
\value{
This function invisibly returns NULL. It is used for its side effect of printing to the console or a file.
}
\description{
This function is a wrapper around the base R 'cat' function, with the default separator set to an empty string.
}
\examples{
cat0("Hello", "World") # prints "HelloWorld"
cat0("Hello", "World", sep = " ") # prints "Hello World"
cat0("Line1", "Line2", sep = "\n") # prints "Line1" and "Line2" on separate lines
}
